requirements.txt
pika
supabase
python-dotenv
requests
fastapi[all]